+++ frontend/catalyst/utils/toml.py
@@ -16,9 +16,10 @@ Module for abstracting which toml_load to use.
 """
 
 import importlib.util
+from dataclasses import dataclass
 from functools import reduce
 from itertools import repeat
-from typing import Any, Dict, List
+from typing import Any, Dict, List, Set
 
 from catalyst.utils.exceptions import CompileError
 
@@ -40,9 +41,11 @@ if tomllib:  # pragma: nocover
     from tomllib import load as toml_load
 
     TOMLDocument = Any
+    TOMLException = Exception
 else:  # pragma: nocover
     from tomlkit import TOMLDocument
     from tomlkit import load as toml_load
+    from tomlkit.exceptions import TOMLKitError as TOMLException
 
 
 def read_toml_file(toml_file: str) -> TOMLDocument:
@@ -52,25 +55,66 @@ def read_toml_file(toml_file: str) -> TOMLDocument:
     return config
 
 
-def check_mid_circuit_measurement_flag(config: TOMLDocument) -> bool:
-    """Check the global mid-circuit measurement flag"""
-    return bool(config.get("compilation", {}).get("mid_circuit_measurement", False))
+@dataclass
+class OperationProperties:
+    """Capabilities of a single operation"""
+
+    invertible: bool
+    controllable: bool
+    differentiable: bool
 
 
-def check_adjoint_flag(config: TOMLDocument, shots_present: bool) -> bool:
-    """Check the global adjoint flag for toml schema 1. For newer schemas the adjoint flag is
-    defined to be set if all native gates are inverible"""
-    schema = int(config["schema"])
-    if schema == 1:
-        return bool(config.get("compilation", {}).get("quantum_adjoint", False))
+def intersect_properties(a: OperationProperties, b: OperationProperties) -> OperationProperties:
+    """Calculate the intersection of OperationProperties"""
+    return OperationProperties(
+        invertible=a.invertible and b.invertible,
+        controllable=a.controllable and b.controllable,
+        differentiable=a.differentiable and b.differentiable,
+    )
 
-    elif schema == 2:
-        return all(
-            "invertible" in v.get("properties", {})
-            for g, v in get_native_gates(config, shots_present).items()
-        )
 
-    raise CompileError("quantum_adjoint flag is not supported in TOMLs schema >= 3")
+@dataclass
+class DeviceCapabilities:  # pylint: disable=too-many-instance-attributes
+    """Quantum device capabilities"""
+
+    native_ops: Dict[str, OperationProperties]
+    to_decomp_ops: Dict[str, OperationProperties]
+    to_matrix_ops: Dict[str, OperationProperties]
+    native_obs: Dict[str, OperationProperties]
+    measurement_processes: Set[str]
+    mid_circuit_measurement_flag: bool
+    runtime_code_generation_flag: bool
+    dynamic_qubit_management_flag: bool
+
+
+def intersect_operations(
+    a: Dict[str, OperationProperties], b: Dict[str, OperationProperties]
+) -> Dict[str, OperationProperties]:
+    """Intersects two sets of oepration properties"""
+    return {k: intersect_properties(a[k], b[k]) for k in (a.keys() & b.keys())}
+
+
+def pennylane_operation_set(config_ops: Dict[str, OperationProperties]) -> Set[str]:
+    """Returns a config section into a set of strings using PennyLane syntax"""
+    ops = set()
+    # Back-mapping from class names to string names
+    for g, props in config_ops.items():
+        ops.update({g})
+        if props.controllable:
+            ops.update({f"C({g})"})
+    return ops
+
+
+@dataclass
+class ProgramFeatures:
+    """Program features, obtained from the user"""
+
+    shots_present: bool
+
+
+def check_compilation_flag(config: TOMLDocument, flag_name: str) -> bool:
+    """Checks the flag in the toml document 'compilation' section."""
+    return bool(config.get("compilation", {}).get(flag_name, False))
 
 
 def check_quantum_control_flag(config: TOMLDocument) -> bool:
@@ -82,13 +126,18 @@ def check_quantum_control_flag(config: TOMLDocument) -> bool:
     raise CompileError("quantum_control flag is not supported in TOMLs schema >= 2")
 
 
-def get_gates(config: TOMLDocument, path: List[str], shots_present: bool) -> Dict[str, dict]:
-    """Read the toml config section specified by `path`. Filters-out gates which don't match
-    condition. For now the only condition we support is `shots_present`."""
+def parse_toml_section(
+    config: TOMLDocument, path: List[str], program_features: ProgramFeatures
+) -> Dict[str, dict]:
+    """Parses the section of toml config file specified by `path`. Filters-out gates which don't
+    match condition. For now the only condition we support is `shots_present`."""
     gates = {}
     analytic = "analytic"
     finiteshots = "finiteshots"
-    iterable = reduce(lambda x, y: x[y], path, config)
+    try:
+        iterable = reduce(lambda x, y: x[y], path, config)
+    except TOMLException as _:  # pylint: disable=broad-exception-caught
+        return {}
     gen = iterable.items() if hasattr(iterable, "items") else zip(iterable, repeat({}))
     for g, values in gen:
         unknown_attrs = set(values) - {"condition", "properties"}
@@ -103,6 +152,7 @@ def get_gates(config: TOMLDocument, path: List[str], shots_present: bool) -> Dic
                 f"Configuration for gate '{str(g)}' has unknown properties: {list(unknown_props)}"
             )
         if "condition" in values:
+            # TODO: do not filter here. Parse the condition and then filter on demand instead.
             conditions = values["condition"]
             unknown_conditions = set(conditions) - {analytic, finiteshots}
             if len(unknown_conditions) > 0:
@@ -115,33 +165,50 @@ def get_gates(config: TOMLDocument, path: List[str], shots_present: bool) -> Dic
                     f"Configuration for gate '{g}' can not contain both "
                     f"`{finiteshots}` and `{analytic}` conditions simultaniosly"
                 )
-            if analytic in conditions and not shots_present:
+            if analytic in conditions and not program_features.shots_present:
                 gates[g] = values
-            elif finiteshots in conditions and shots_present:
+            elif finiteshots in conditions and program_features.shots_present:
                 gates[g] = values
         else:
             gates[g] = values
     return gates
 
 
-def get_observables(config: TOMLDocument, shots_present: bool) -> Dict[str, dict]:
+def get_observables(config: TOMLDocument, program_features: ProgramFeatures) -> Dict[str, dict]:
     """Override the set of supported observables."""
-    return get_gates(config, ["operators", "observables"], shots_present)
+    return parse_toml_section(config, ["operators", "observables"], program_features)
 
 
-def get_native_gates(config: TOMLDocument, shots_present: bool) -> Dict[str, dict]:
-    """Get the gates from the `native` section of the config."""
+def get_measurement_processes(
+    config: TOMLDocument, program_features: ProgramFeatures
+) -> Dict[str, dict]:
+    """Get the measurements processes from the `native` section of the config."""
 
     schema = int(config["schema"])
     if schema == 1:
-        return get_gates(config, ["operators", "gates", 0, "native"], shots_present)
-    elif schema == 2:
-        return get_gates(config, ["operators", "gates", "native"], shots_present)
+        shots_string = "finiteshots" if program_features.shots_present else "exactshots"
+        return parse_toml_section(config, ["measurement_processes", shots_string], program_features)
+    if schema == 2:
+        return parse_toml_section(config, ["measurement_processes"], program_features)
 
     raise CompileError(f"Unsupported config schema {schema}")
 
 
-def get_decomposable_gates(config: TOMLDocument, shots_present: bool) -> Dict[str, dict]:
+def get_native_ops(config: TOMLDocument, program_features: ProgramFeatures) -> Dict[str, dict]:
+    """Get the gates from the `native` section of the config."""
+
+    schema = int(config["schema"])
+    if schema == 1:
+        return parse_toml_section(config, ["operators", "gates", 0, "native"], program_features)
+    elif schema == 2:
+        return parse_toml_section(config, ["operators", "gates", "native"], program_features)
+
+    raise CompileError(f"Unsupported config schema {schema}")
+
+
+def get_decomposable_gates(
+    config: TOMLDocument, program_features: ProgramFeatures
+) -> Dict[str, dict]:
     """Get gates that will be decomposed according to PL's decomposition rules.
 
     Args:
@@ -149,14 +216,16 @@ def get_decomposable_gates(config: TOMLDocument, shots_present: bool) -> Dict[st
     """
     schema = int(config["schema"])
     if schema == 1:
-        return get_gates(config, ["operators", "gates", 0, "decomp"], shots_present)
+        return parse_toml_section(config, ["operators", "gates", 0, "decomp"], program_features)
     elif schema == 2:
-        return get_gates(config, ["operators", "gates", "decomp"], shots_present)
+        return parse_toml_section(config, ["operators", "gates", "decomp"], program_features)
 
     raise CompileError(f"Unsupported config schema {schema}")
 
 
-def get_matrix_decomposable_gates(config: TOMLDocument, shots_present: bool) -> Dict[str, dict]:
+def get_matrix_decomposable_gates(
+    config: TOMLDocument, program_features: ProgramFeatures
+) -> Dict[str, dict]:
     """Get gates that will be decomposed to QubitUnitary.
 
     Args:
@@ -164,8 +233,144 @@ def get_matrix_decomposable_gates(config: TOMLDocument, shots_present: bool) -> 
     """
     schema = int(config["schema"])
     if schema == 1:
-        return get_gates(config, ["operators", "gates", 0, "matrix"], shots_present)
+        return parse_toml_section(config, ["operators", "gates", 0, "matrix"], program_features)
     elif schema == 2:
-        return get_gates(config, ["operators", "gates", "matrix"], shots_present)
+        return parse_toml_section(config, ["operators", "gates", "matrix"], program_features)
 
     raise CompileError(f"Unsupported config schema {schema}")
+
+
+def get_operation_properties(config_props: dict) -> OperationProperties:
+    """Load operation properties from config"""
+    properties = config_props.get("properties", {})
+    return OperationProperties(
+        invertible="invertible" in properties,
+        controllable="controllable" in properties,
+        differentiable="differentiable" in properties,
+    )
+
+
+def patch_schema1_collections(
+    config, device_name, native_gate_props, matrix_decomp_props, decomp_props, observable_props
+):  # pylint: disable=too-many-arguments, too-many-branches
+    """For old schema1 config files we deduce some information which was not explicitly encoded."""
+
+    # TODO: remove after PR #642 is merged in lightning
+    # NOTE: we mark GlobalPhase as controllables even if `quantum_control` flag is False. This
+    # is what actual device reports.
+    if device_name == "lightning.kokkos":  # pragma: nocover
+        native_gate_props["GlobalPhase"] = OperationProperties(
+            invertible=False, controllable=True, differentiable=True
+        )
+
+    # TODO: remove after PR #642 is merged in lightning
+    if device_name == "lightning.kokkos":  # pragma: nocover
+        observable_props["Projector"] = OperationProperties(
+            invertible=False, controllable=False, differentiable=False
+        )
+
+    # The deduction logic is the following:
+    # * Most of the gates have their `C(Gate)` controlled counterparts.
+    # * Some gates have to be decomposed if controlled version is used. Typically these are
+    #   gates which are already controlled but have well-known names.
+    # * Few gates, like `QubitUnitary`, have separate classes for their controlled versions.
+    gates_to_be_decomposed_if_controlled = [
+        "Identity",
+        "CNOT",
+        "CY",
+        "CZ",
+        "CSWAP",
+        "CRX",
+        "CRY",
+        "CRZ",
+        "CRot",
+        "ControlledPhaseShift",
+        "QubitUnitary",
+        "ControlledQubitUnitary",
+        "Toffoli",
+    ]
+
+    supports_controlled = check_quantum_control_flag(config)
+    if supports_controlled:
+        # Add ControlledQubitUnitary as a controlled version of QubitUnitary
+        if "QubitUnitary" in native_gate_props:
+            native_gate_props["ControlledQubitUnitary"] = OperationProperties(
+                invertible=False, controllable=False, differentiable=True
+            )
+        # By default, enable the `C(gate)` version for most `gates`.
+        for op, props in native_gate_props.items():
+            props.controllable = op not in gates_to_be_decomposed_if_controlled
+
+    supports_adjoint = check_compilation_flag(config, "quantum_adjoint")
+    if supports_adjoint:
+        # Makr all gates as invertibles
+        for props in native_gate_props.values():
+            props.invertible = True
+
+    # For toml schema 1 configs, the following condition is possible: (1) `QubitUnitary` gate is
+    # supported, (2) native quantum control flag is enabled and (3) `ControlledQubitUnitary` is
+    # listed in either matrix or decomposable sections. This is a contradiction, because
+    # condition (1) means that `ControlledQubitUnitary` is also in the native set. We solve it
+    # here by applying a fixup.
+    # TODO: remove after PR #642 is merged in lightning
+    if "ControlledQubitUnitary" in native_gate_props:  # pragma: nocover
+        if "ControlledQubitUnitary" in matrix_decomp_props:
+            matrix_decomp_props.pop("ControlledQubitUnitary")
+        if "ControlledQubitUnitary" in decomp_props:
+            decomp_props.pop("ControlledQubitUnitary")
+
+    # Fix a bug in device toml schema 1
+    if "ControlledPhaseShift" in native_gate_props:  # pragma: nocover
+        if "ControlledPhaseShift" in matrix_decomp_props:
+            matrix_decomp_props.pop("ControlledPhaseShift")
+        if "ControlledPhaseShift" in decomp_props:
+            decomp_props.pop("ControlledPhaseShift")
+
+
+def get_device_capabilities(
+    config: TOMLDocument, program_features: ProgramFeatures, device_name: str
+) -> DeviceCapabilities:
+    """Load TOML document into the DeviceCapabilities structure"""
+
+    schema = int(config["schema"])
+
+    native_gate_props = {}
+    for g, props in get_native_ops(config, program_features).items():
+        native_gate_props[g] = get_operation_properties(props)
+
+    matrix_decomp_props = {}
+    for g, props in get_matrix_decomposable_gates(config, program_features).items():
+        matrix_decomp_props[g] = get_operation_properties(props)
+
+    decomp_props = {}
+    for g, props in get_decomposable_gates(config, program_features).items():
+        decomp_props[g] = get_operation_properties(props)
+
+    observable_props = {}
+    for g, props in get_observables(config, program_features).items():
+        observable_props[g] = get_operation_properties(props)
+
+    measurements_props = set()
+    for g, props in get_measurement_processes(config, program_features).items():
+        measurements_props.add(g)
+
+    if schema == 1:
+        patch_schema1_collections(
+            config,
+            device_name,
+            native_gate_props,
+            matrix_decomp_props,
+            decomp_props,
+            observable_props,
+        )
+
+    return DeviceCapabilities(
+        native_ops=native_gate_props,
+        to_decomp_ops=decomp_props,
+        to_matrix_ops=matrix_decomp_props,
+        native_obs=observable_props,
+        measurement_processes=measurements_props,
+        mid_circuit_measurement_flag=check_compilation_flag(config, "mid_circuit_measurement"),
+        runtime_code_generation_flag=check_compilation_flag(config, "runtime_code_generation"),
+        dynamic_qubit_management_flag=check_compilation_flag(config, "dynamic_qubit_management"),
+    )