+++ frontend/test/pytest/test_autograph.py
@@ -38,7 +38,12 @@ from catalyst import (
     qjit,
     vjp,
 )
-from catalyst.autograph import TRANSFORMER, AutoGraphError, autograph_source
+from catalyst.autograph import (
+    TRANSFORMER,
+    AutoGraphError,
+    autograph_source,
+    disable_autograph,
+)
 
 check_cache = TRANSFORMER.has_cache
 
@@ -67,7 +72,6 @@ class Failing:
         return self.ref
 
 
-@pytest.mark.tf
 class TestSourceCodeInfo:
     """Unit tests for exception utilities that retrieves traceback information for the original
     source code."""
@@ -149,7 +153,6 @@ class TestSourceCodeInfo:
                 assert e.args == ("Test failure",)
 
 
-@pytest.mark.tf
 class TestIntegration:
     """Test that the autograph transformations trigger correctly in different settings."""
 
@@ -394,7 +397,6 @@ class TestIntegration:
         assert np.allclose(fn(3)[1], tuple([jnp.array(2.0), jnp.array(6.0)]))
 
 
-@pytest.mark.tf
 class TestCodePrinting:
     """Test that the transformed source code can be printed in different settings."""
 
@@ -524,7 +526,6 @@ class TestCodePrinting:
         assert autograph_source(inner)
 
 
-@pytest.mark.tf
 class TestConditionals:
     """Test that the autograph transformations produce correct results on conditionals.
     These tests are adapted from the test_conditionals.TestCond class of tests."""
@@ -690,7 +691,6 @@ class TestConditionals:
             qjit(autograph=True)(f)
 
 
-@pytest.mark.tf
 class TestForLoops:
     """Test that the autograph transformations produce correct results on for loops."""
 
@@ -1244,7 +1244,6 @@ class TestForLoops:
         assert f() == 9
 
 
-@pytest.mark.tf
 class TestWhileLoops:
     """Test that the autograph transformations produce correct results on while loops."""
 
@@ -1427,7 +1426,6 @@ class TestWhileLoops:
             qjit(autograph=True)(f)
 
 
-@pytest.mark.tf
 @pytest.mark.parametrize(
     "execution_context", (lambda fn: fn, qml.qnode(qml.device("lightning.qubit", wires=1)))
 )
@@ -1537,7 +1535,6 @@ class TestFallback:
         assert results[1] == (2) * 1 * 2 * 3  # i = range(1, 4)
 
 
-@pytest.mark.tf
 class TestLogicalOps:
     """Test logical operations: and, or, not"""
 
@@ -1612,7 +1609,6 @@ class TestLogicalOps:
             assert qjit(autograph=True)(lambda d: d or s)(d) == (d or s)
 
 
-@pytest.mark.tf
 class TestMixed:
     """Test a mix of supported autograph conversions and Catalyst control flow."""
 
@@ -1716,6 +1712,30 @@ class TestMixed:
 
         assert f(0.4) == 1
         assert f(0.1) == 3
+
+
+class TestDisableAutograph:
+    """Test ways of disabling autograph conversion"""
+
+    def test_disable_autograph(self):
+        """Test disabling autograph with decorator."""
+
+        @disable_autograph
+        def f():
+            x = 6
+            if x > 5:
+                y = x**2
+            else:
+                y = x**3
+            return y
+
+        @qjit(autograph=True)
+        def g(x: float, n: int):
+            for _ in range(n):
+                x = x + f()
+            return x
+
+        assert g(0.4, 6) == 216.4
 
 
 if __name__ == "__main__":