+++ frontend/test/pytest/test_braket_remote_devices.py
@@ -20,9 +20,7 @@ import pytest
 
 from catalyst import grad, qjit
 
-pytest.importorskip("braket")
-
-from braket.devices import Devices  # pylint: disable=wrong-import-position
+braket = pytest.importorskip("braket")
 
 try:
     qml.device(
@@ -53,7 +51,7 @@ class TestBraketS3Bucket:
             ),
             qml.device(
                 "braket.aws.qubit",
-                device_arn=Devices.Amazon.SV1,
+                device_arn=braket.devices.Devices.Amazon.SV1,
                 s3_destination_folder=("my-bucket", "my-prefix"),
                 wires=2,
             ),
@@ -88,7 +86,7 @@ class TestBraketGates:
             ),
             qml.device(
                 "braket.aws.qubit",
-                device_arn=Devices.Amazon.SV1,
+                device_arn=braket.devices.Devices.Amazon.SV1,
                 wires=3,
             ),
         ],
@@ -139,7 +137,7 @@ class TestBraketGates:
             ),
             qml.device(
                 "braket.aws.qubit",
-                device_arn=Devices.Amazon.SV1,
+                device_arn=braket.devices.Devices.Amazon.SV1,
                 wires=3,
             ),
         ],
@@ -214,11 +212,11 @@ class TestBraketSample:
             qml.RX(x, wires=0)
             return qml.sample()
 
-        expected = np.array([[0.0]] * 1000)
+        expected = np.array([[0]] * 1000)
         observed = sample_1qbit(0.0)
         assert np.array_equal(observed, expected)
 
-        expected = np.array([[1.0]] * 1000)
+        expected = np.array([[1]] * 1000)
         observed = sample_1qbit(np.pi)
         assert np.array_equal(observed, expected)
 
@@ -243,10 +241,10 @@ class TestBraketSample:
             qml.RY(x, wires=1)
             return qml.sample()
 
-        expected = np.array([[0.0, 0.0]] * 1000)
+        expected = np.array([[0, 0]] * 1000)
         observed = sample_2qbits(0.0)
         assert np.array_equal(observed, expected)
-        expected = np.array([[1.0, 1.0]] * 1000)
+        expected = np.array([[1, 1]] * 1000)
         observed = sample_2qbits(np.pi)
         assert np.array_equal(observed, expected)
 
@@ -718,9 +716,11 @@ class TestBraketMeasurementsProcess:
 
         # qml.sample
         assert result[0].shape[0] == 100
+        assert result[0].dtype == np.int64
 
         # qml.counts
         assert sum(result[1][1]) == 100
+        assert result[1][0].dtype == np.int64
 
         # qml.probs
         assert result[2][0] > result[2][1]