+++ mlir/lib/Catalyst/Transforms/RegisterAllPasses.cpp
@@ -39,4 +39,5 @@ void catalyst::registerAllCatalystPasses()
     mlir::registerPass(catalyst::createAddExceptionHandlingPass);
     mlir::registerPass(catalyst::createGEPInboundsPass);
     mlir::registerPass(catalyst::createRemoveChainedSelfInversePass);
+    mlir::registerPass(catalyst::createAnnotateFunctionPass);
 }